retail_store/carrefour/requirements.txt
playwright
playwright-stealth
psycopg2-binary
SQLAlchemy
python-dotenv
